refactor: Fallow - break import cycle to reduce change cascade risk
