+++ packages/common/src/extensions/slickGroupItemMetadataProvider.ts
@@ -1,12 +1,8 @@
 import { createDomElement, extend, isHtml } from '@slickgrid-universal/utils';
-import {
-  applyHtmlToElement,
-  SlickEventHandler,
-  SlickGroup,
-  type SlickDataView,
-  type SlickEventData,
-  type SlickGrid,
-} from '../core/index.js';
+import { SlickEventHandler, SlickGroup, type SlickEventData } from '../core/slickCore.js';
+import type { SlickDataView } from '../core/slickDataview.js';
+import type { SlickGrid } from '../core/slickGrid.js';
+import { applyHtmlToElement } from '../core/utils.js';
 import type {
   Column,
   GridOption,
